+++ aegis_grpc/python_client/aegis_grpc_client/__init__.py
@@ -1,0 +1,3 @@
+from .grpc_client import AegisRobotClient
+
+__all__ = ["AegisRobotClient"]