+++ aegis_grpc/python_client/test/import_test.py
@@ -1,0 +1,6 @@
+def test_import():
+    import aegis_grpc_client  # noqa: F401
+
+
+def test_import_AegisRobotClient():
+    from aegis_grpc_client import AegisRobotClient  # noqa: F401